Move harness and server to render folder, simplify render code to be sepcific.
